Now core and gui is working.
There are probably some bugs in JS side due to some changes on the json-structure. We need to check Conditions since they need isa also on ValueHolder. Maybe change the Condition structure on scala side instead...

All services in extension is left to upgrade. Also some logic needs to be migrated to the new structure.

When everything is working, I will create a new service structure.